refactor entityreducer into two parts
