+++ lib/src/actions/entity-cache-action.ts
@@ -5,6 +5,7 @@ import { Action } from '@ngrx/store';
 
 import { EntityCache } from '../reducers/entity-cache';
 import { EntityActionOptions } from '../actions/entity-action';
+import { MergeStrategy } from '../actions/merge-strategy';
 
 export enum EntityCacheAction {
   MERGE_QUERY_SET = 'ngrx-data/entity-cache/merge-query-set',
@@ -20,10 +21,25 @@ export interface EntityCacheQuerySet {
  * that returned entities from multiple collections.
  * Corresponding entity cache reducer should add and update all collections
  * at the same time, before any selectors$ observables emit.
+ * @param querySet The result of the query in the form of a map of entity collections.
+ * These are the entity data to merge into the respective collections.
+ * @param mergeStrategy How to merge a queried entity when it is already in the collection.
+ * The default is MergeStrategy.PreserveChanges
  */
 export class MergeQuerySet implements Action {
+  readonly payload: {
+    querySet: EntityCacheQuerySet;
+    mergeStrategy?: MergeStrategy;
+  };
+
   readonly type = EntityCacheAction.MERGE_QUERY_SET;
-  constructor(public readonly payload: EntityCacheQuerySet, options?: EntityActionOptions) {}
+
+  constructor(querySet: EntityCacheQuerySet, mergeStrategy?: MergeStrategy) {
+    this.payload = {
+      querySet,
+      mergeStrategy: mergeStrategy === null ? MergeStrategy.PreserveChanges : mergeStrategy
+    };
+  }
 }
 
 /**
