+++ lib/src/reducers/constants.ts
@@ -3,20 +3,9 @@ import { Action, ActionReducer, MetaReducer } from '@ngrx/store';
 import { EntityCache } from './entity-cache';
 
 export const ENTITY_CACHE_NAME = 'entityCache';
-export const ENTITY_CACHE_NAME_TOKEN = new InjectionToken<string>(
-  'ngrx-data/entity-cache-name'
-);
+export const ENTITY_CACHE_NAME_TOKEN = new InjectionToken<string>('ngrx-data/entity-cache-name');
 
-export const ENTITY_CACHE_META_REDUCERS = new InjectionToken<
-  MetaReducer<any, any>[]
->('ngrx-data/entity-cache-meta-reducers');
-export const ENTITY_COLLECTION_META_REDUCERS = new InjectionToken<
-  MetaReducer<any, any>[]
->('ngrx-data/entity-collection-meta-reducers');
-export const ENTITY_CACHE_REDUCER = new InjectionToken<
-  ActionReducer<EntityCache>
->('ngrx-data/entity-reducer');
+export const ENTITY_CACHE_META_REDUCERS = new InjectionToken<MetaReducer<any, any>[]>('ngrx-data/entity-cache-meta-reducers');
+export const ENTITY_COLLECTION_META_REDUCERS = new InjectionToken<MetaReducer<any, any>[]>('ngrx-data/entity-collection-meta-reducers');
 
-export const INITIAL_ENTITY_CACHE_STATE = new InjectionToken<
-  EntityCache | (() => EntityCache)
->('ngrx-data/initial-entity-cache-state');
+export const INITIAL_ENTITY_CACHE_STATE = new InjectionToken<EntityCache | (() => EntityCache)>('ngrx-data/initial-entity-cache-state');
